started on client and admin login logic, used prettier

+++ src/Components/Manage.jsx
@@ -93,7 +93,7 @@ const Manage = ({ users, setUsers }) => {
             return (
               <li key={idx}>
                 <div className="name">
-                  {parseInt(user.amount) >= 1000000 &&
+                  {parseFloat(user.amount) >= 1000000 &&
                     user.status === "ACTIVE" && <RiVipCrownFill />}
                   {user.firstName + " " + user.lastName ===
                     "Jhowell Santiago" && <RiStarFill />}
